Migrate logs to S3 from the middle
When migrating logs to S3, save the last migrated log id at DB and in the next time migrate logs from the next id.
Before we stop the Drone for its upgrade, we can migrate almost logs online so we can really decrease the downtime for the upgrade.
It takes a very long time to migrate logs to S3 if the logs table is huge, so this feature is very important.